+++ base_map/base_map_manager.c
@@ -18,8 +18,27 @@
 #include "server.h"
 #include "player.h"
 #include "sdl.h"
+#include "data.h"
 #include "base_map.h"
 #include "game_info.h"
+
+// wrapper to init an SDL_Rect
+SDL_Rect init_rect(int x, int y, int w, int z)
+{
+  SDL_Rect temp = {x, y, w, z};
+
+  return (temp);
+}
+
+t_map init_t_map(SDL_Rect src, SDL_Rect dest, texture_type type)
+{
+  t_map map;
+
+  map.src_rect = src;
+  map.dest_rect = dest;
+  map.texture_type = type;
+  return (map);
+}
 
 void *draw_all(void *arg)
 {
