+++ includes/draw_base_map.h
@@ -1,0 +1,8 @@
+#ifndef _DRAW_BASE_MAP_H_
+#define _DRAW_BASE_MAP_H_
+
+void *draw_map_model(void *arg);
+void *draw_timer(void *arg);
+void draw_pannel(void *arg);
+
+#endif /* !_DRAW_BASE_MAP_H_ */
