+++ includes/start_map.h
@@ -1,0 +1,6 @@
+#ifndef _START_MAP_H_
+#define _START_MAP_H_
+
+int start_map(t_sdl *sdl, int socket, t_player_request *player_request);
+
+#endif _START_MAP_H_
